+++ site/theme/default/App.vue
@@ -127,7 +127,7 @@ export default {
     },
     changeLang () {
       const lang = this.lang === 'zh-CN' ? 'en-US' : 'zh-CN'
-      localStore('MAND_MOBILE_LANG', lang)
+      // localStore('AL_MOBILE_LANG', lang)
       location.href = location.href.replace(this.lang, lang)
 
       if (~location.href.indexOf('home')) {