+++ CHANGELOG.md
@@ -3,11 +3,22 @@ title: 更新日志
 toc: hidden
 ---
 
+### 0.0.25
+
+`2020-05-06`
+
+- Feat
+  - `Dialog`属性`title`、`btn`，由v-text修改为v-html，因此0.0.25版本可以使用String或者vnode
+- Fix
+  - `head`修复头部样式污染n22-icon样式问题
+  - `switch-item`修复changeData改变回掉事件中，第二个参数value值错误问题
+
 ### 0.0.24
 
 `2020-05-06`
 
 - Fix
+  - `content`修复tab中isMescrollUp
   - `switch-item`修复没有options时，true跟false的boolearn值回显赋值问题（mounted赋值错误）
   - `switch-item`修复changeData改变回掉事件中，第二个参数value值错误问题